+++ Annoy-o-Bot/CallbackHandler.cs
@@ -119,26 +119,6 @@ namespace Annoy_o_Bot
         public string Path { get; set; }
     }
 
-    public class Reminder
-    {
-        public string Title { get; set; }
-        public string Message { get; set; }
-        public string Assignee { get; set; }
-        public DateTime Date { get; set; }
-        public Interval Interval { get; set; }
-        public int? IntervalStep { get; set; }
-    }
-
-    public enum Interval
-    {
-        Once = 0,
-        Daily = 1,
-        Weekly = 2,
-        Monthly = 3,
-        Yearly = 4
-    }
-
-
     //TODO: support projects
 
 }
